fix(sonar): resolve ~170 SonarCloud issues across multiple rules
S1905 - Remove 70 unnecessary casts (IEnumerable<T> and int)
S1172 - Fix 20 unused method parameters (discard or remove)
S1481 - Fix 9 unused local variables (remove or discard)
S1066 - Merge 8 nested if statements
S3267 - Simplify 8 loops with LINQ (Where, Select, Any)
S6580 - Add CultureInfo.InvariantCulture to 12 DateTime.Parse calls
S2589 - Remove 2 always-true/false conditions
S2583 - Fix 2 unreachable code bugs (committed earlier)
S8264 - Move workflow permissions to job level (committed earlier)

Refs #76

+++ src/Encina.ADO.MySQL/Tenancy/TenantAwareFunctionalRepositoryADO.cs
@@ -566,7 +566,7 @@ public sealed class TenantAwareFunctionalRepositoryADO<TEntity, TId> : IFunction
             if (entity is IVersionedEntity versionedEntity)
             {
                 var originalVersion = versionedEntity.Version;
-                versionedEntity.Version = (int)(originalVersion + 1);
+                versionedEntity.Version = originalVersion + 1;
 
                 var versionedSql = BuildVersionedUpdateSql();
 
@@ -824,7 +824,7 @@ public sealed class TenantAwareFunctionalRepositoryADO<TEntity, TId> : IFunction
                     if (entity is IVersionedEntity versionedEntity)
                     {
                         var originalVersion = versionedEntity.Version;
-                        versionedEntity.Version = (int)(originalVersion + 1);
+                        versionedEntity.Version = originalVersion + 1;
 
                         using var command = _connection.CreateCommand();
                         command.CommandText = versionedSql;
