+++ src/Encina.MongoDB/SoftDelete/SoftDeletableFunctionalRepositoryMongoDB.cs
@@ -666,7 +666,7 @@ public sealed class SoftDeletableFunctionalRepositoryMongoDB<TEntity, TId>
                 if (entity is IVersionedEntity versionedEntity)
                 {
                     originalVersions[entity] = versionedEntity.Version;
-                    versionedEntity.Version = (int)(versionedEntity.Version + 1);
+                    versionedEntity.Version = versionedEntity.Version + 1;
                 }
                 else if (entity is IVersioned versioned)
                 {
